🔨 Refactor(#28): 네이밍, 타입, 파일이름 수정
