--- src/features/part/service/parsePartData.ts
@@ -1,8 +1,0 @@
-import { Part } from '../types';
-
-export const parsePartData = (part: Record<string, any>): Omit<Part, 'id'> => {
-  return {
-    name: part.partName.toString(),
-    sectionId: Number(part.sectionId),
-  };
-};
