+++ src/features/part/service/utils.ts
@@ -1,0 +1,10 @@
+import { Part } from '../types';
+
+export const parsePartData = (
+  formData: Record<string, FormDataEntryValue>
+): Omit<Part, 'id'> => {
+  return {
+    name: formData.partName.toString(),
+    sectionId: Number(formData.sectionId),
+  };
+};
