+++ src/features/part/ui/PartForm.tsx
@@ -1,6 +1,6 @@
 import { Alert, Button, FloatingLabel, Form, Row } from 'react-bootstrap';
 import partsQueries from '../queries';
-import { parsePartData } from '../service/parsePartData';
+import { parsePartData } from '../service/utils';
 import { useState } from 'react';
 import sectionsQueries from '../../section/queries';
